requirements.txt
phidata
phidata[all]
agno
python-dotenv
duckduckgo-search
sqlalchemy
groq
pymysql
yfinance
wikipedia
ddgs